refactor(LottoResult): LottoResult domain으로 이동
로또 결과를 계산도 해주는 LottoResult는 domain 패키지에 있는게 맞다고 생각하여 이동시켰습니다.
