+++ src/test/java/lotto/data/LottoResultTest.java
@@ -7,6 +7,7 @@ import java.math.RoundingMode;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Stream;
+import lotto.domain.LottoResult;
 import lotto.message.LottoPrize;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
